refactor(workers): isolate error formatting
