+++ src/gateway/worker-environments/service-validation.ts
@@ -1,7 +1,4 @@
 import { isRecord } from "@openclaw/normalization-core/record-coerce";
-import { truncateUtf16Safe } from "@openclaw/normalization-core/utf16-slice";
-import { formatErrorMessage } from "../../infra/errors.js";
-import { redactSensitiveText } from "../../logging/redact.js";
 import type {
   WorkerDesktopEndpoint,
   WorkerLease,
@@ -48,11 +45,4 @@ export function requireWorkerLease(value: unknown): WorkerLease {
       ? {}
       : { desktop: normalizeWorkerDesktopEndpoint(value.desktop as WorkerDesktopEndpoint) }),
   };
-}
-
-export function boundedWorkerError(error: unknown): string {
-  const redacted = redactSensitiveText(formatErrorMessage(error), { mode: "tools" })
-    .replace(/\s+/g, " ")
-    .trim();
-  return truncateUtf16Safe(redacted || "unknown error", 1_024);
 }
